requirements.txt
# requirements.txt
Flask>=2.0
openai>=1.0
python-dotenv>=0.19
werkzeug>=2.0
Flask-Session>=0.4
Flask-SSE>=0.2.1
Markdown>=3.3
redis>=4.0 # Add redis client library

# For production deployment (example)
# gunicorn